+++ zebra-state/src/lib.rs
@@ -75,7 +75,7 @@ pub use service::finalized_state::{ReadDisk, TypedColumnFamily, WriteTypedBatch}
 
 pub use service::{
     finalized_state::{DiskWriteBatch, FromDisk, IntoDisk, WriteDisk, ZebraDb},
-    ReadStateService,
+    ReadStateService, StateInfo,
 };
 
 // Allow use in external tests